Feat: #65 - review page 컴포넌트 분리
